+++ src/main/kotlin/dev/frozenmilk/dairy/calcified/hardware/motor/CalcifiedMotor.kt
@@ -13,14 +13,14 @@ import dev.frozenmilk.dairy.core.util.current.CurrentUnits
 import kotlin.math.abs
 import kotlin.math.roundToInt
 
-class CalcifiedMotor internal constructor(val module: CalcifiedModule, val port: Byte) : ComplexMotor {
+class CalcifiedMotor internal constructor(val module: CalcifiedModule, val port: Int) : ComplexMotor {
 	override var direction = Direction.FORWARD
 	override var cachingTolerance = 0.005
 	override var enabled = true
 		set(value) {
 			if (field != value) {
 				// sends the command to change the enable state
-				LynxSetMotorChannelEnableCommand(module.lynxModule, port.toInt(), value).send()
+				LynxSetMotorChannelEnableCommand(module.lynxModule, port, value).send()
 				field = value
 			}
 		}
@@ -29,7 +29,7 @@ class CalcifiedMotor internal constructor(val module: CalcifiedModule, val port:
 		set(value) {
 			if (field != value) {
 				// sets the command to change the 0 power behaviour
-				LynxSetMotorChannelModeCommand(module.lynxModule, port.toInt(), DcMotor.RunMode.RUN_WITHOUT_ENCODER, value.wrapping)
+				LynxSetMotorChannelModeCommand(module.lynxModule, port, DcMotor.RunMode.RUN_WITHOUT_ENCODER, value.wrapping)
 				field = value
 			}
 		}
@@ -40,27 +40,27 @@ class CalcifiedMotor internal constructor(val module: CalcifiedModule, val port:
 			if (!enabled) return
 			val correctedValue = value.coerceIn(-1.0, 1.0) * direction.multiplier
 			if (abs(field - correctedValue) >= cachingTolerance || (correctedValue >= 1.0 && field != 1.0) || (correctedValue <= -1.0 && field != -1.0)) {
-				LynxSetMotorConstantPowerCommand(module.lynxModule, port.toInt(), (correctedValue * LynxSetMotorConstantPowerCommand.apiPowerLast).toInt()).send()
+				LynxSetMotorConstantPowerCommand(module.lynxModule, port, (correctedValue * LynxSetMotorConstantPowerCommand.apiPowerLast).toInt()).send()
 				field = correctedValue
 			}
 		}
 
 	override val current: Current
 		get() {
-			val milliAmps = LynxGetADCCommand(module.lynxModule, LynxGetADCCommand.Channel.motorCurrent(port.toInt()), LynxGetADCCommand.Mode.ENGINEERING).sendReceive().value
+			val milliAmps = LynxGetADCCommand(module.lynxModule, LynxGetADCCommand.Channel.motorCurrent(port), LynxGetADCCommand.Mode.ENGINEERING).sendReceive().value
 			return Current(CurrentUnits.MILLI_AMP, milliAmps.toDouble())
 		}
-	override var overCurrentThreshold = Current(CurrentUnits.MILLI_AMP, LynxGetMotorChannelCurrentAlertLevelCommand(module.lynxModule, port.toInt()).sendReceive().currentLimit.toDouble())
+	override var overCurrentThreshold = Current(CurrentUnits.MILLI_AMP, LynxGetMotorChannelCurrentAlertLevelCommand(module.lynxModule, port).sendReceive().currentLimit.toDouble())
 		set(value) {
-			LynxSetMotorChannelCurrentAlertLevelCommand(module.lynxModule, port.toInt(), value.into(CurrentUnits.MILLI_AMP).value.roundToInt()).send()
+			LynxSetMotorChannelCurrentAlertLevelCommand(module.lynxModule, port, value.into(CurrentUnits.MILLI_AMP).value.roundToInt()).send()
 			field = value
 		}
 	override val overCurrent: Boolean
-		get() = module.bulkData.isOverCurrent(port.toInt())
+		get() = module.bulkData.isOverCurrent(port)
 
 	init {
-		LynxSetMotorChannelEnableCommand(module.lynxModule, port.toInt(), true).send()
-		LynxSetMotorChannelModeCommand(module.lynxModule, port.toInt(), DcMotor.RunMode.RUN_WITHOUT_ENCODER, ZeroPowerBehaviour.FLOAT.wrapping)
+		LynxSetMotorChannelEnableCommand(module.lynxModule, port, true).send()
+		LynxSetMotorChannelModeCommand(module.lynxModule, port, DcMotor.RunMode.RUN_WITHOUT_ENCODER, ZeroPowerBehaviour.FLOAT.wrapping)
 	}
 }
